+++ src/flag_gems/runtime/backend/_nvidia/__init__.py
@@ -1,4 +1,4 @@
-from backend_utils import VendorInfoBase  # noqa: E402
+from backend_utils import Autograd, VendorInfoBase  # noqa: E402
 
 from .ops import *  # noqa: F403
 
@@ -7,8 +7,8 @@ vendor_info = VendorInfoBase(
 )
 
 
-def get_register_ops():
-    return (("add.Tensor", add, False),)
+def get_register_op_config():
+    return (("add.Tensor", add, Autograd.disable),)
 
 
 def get_unused_op():
